% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\docType{data}
\name{abstracts}
\alias{abstracts}
\title{Zotero example}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3529 rows and 87 columns.
}
\usage{
data(abstracts)
}
\value{
A dataframe named 'abstracts' with the minimum required variables for making the 'topicspace' package work correctly, namely: ID, documents, title, start.year. It also returns 'participants' dataframe with additional variables categorizing authors, which can be projected into the space of topics.
}
\description{
A dataset of scientific journal abstracts exported directly from Zotero. The
necessary recodes are included here so that it is easier to replicate on your
own Zotero library.
}
\examples{

library(dplyr)
library(purrr)

data(abstracts)
abstracts$nchar <- nchar(abstracts$`Abstract Note`)
abstracts       <- abstracts \%>\% filter(!is.na(nchar)) \%>\% filter(nchar > 500)
nmiss <- abstracts \%>\% map_dbl(~sum(is.na(.x))/length(.x))
abstracts     <- abstracts[, nmiss < 0.90]

# Required columns -----
abstracts$ID           <- abstracts$Key
abstracts$documents     <- abstracts$`Abstract Note`
abstracts$title        <- abstracts$Title
abstracts$start.year   <- abstracts$`Date Added` \%>\% lubridate::year()

# Distinct ----
abstracts    <- abstracts \%>\% distinct(documents, .keep_all = TRUE)
abstracts    <- abstracts \%>\% distinct(DOI, .keep_all = TRUE)
abstracts    <- abstracts \%>\% distinct(title, .keep_all = TRUE)

# Authors ----
authors <- abstracts \%>\% select(ID, Author = Author)
authors <- authors \%>\%
tidyr::separate_rows(Author, sep = ";") \%>\%
mutate(Author = trimws(Author))

}
\keyword{datasets}
