% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{as.sparseMatrix}
\alias{as.sparseMatrix}
\title{Convert simple triplet matrix to sparse matrix}
\usage{
as.sparseMatrix(simple_triplet_matrix_sparse)
}
\arguments{
\item{simple_triplet_matrix_sparse}{A simple triplet matrix in a sparse format.}
}
\value{
A sparse matrix object created from the provided simple triplet matrix.
}
\description{
This function converts a simple triplet matrix in a sparse format to a sparse matrix object. It is implemented in other 'topicspace' functions to convert the 'docterm' document-term matrix to a sparse matrix object before passing it to the LDA model fitting process.
}
\examples{
# Create a simple triplet matrix
stm <- list(
  i = c(1, 2, 3),
  j = c(1, 2, 3),
  v = c(4, 5, 6),
  nrow = 3,
  ncol = 3
)

# Convert the simple triplet matrix to a sparse matrix
sparse_mat <- as.sparseMatrix(stm)
sparse_mat
}
\seealso{
\code{\link{documents.to.topics}}, \code{\link{docterm.to.lda}}, \code{\link{topics.to.mca}}.
}
