% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{documents.to.topics}
\alias{documents.to.topics}
\title{Step 1. Generate topic models from documents}
\usage{
documents.to.topics(
  abstracts,
  label = "",
  lda.k.seq = seq(from = 50, to = 200, by = 50),
  lda.prune.row = 25,
  lda.prune.col = 5,
  filter.terms = NULL
)
}
\arguments{
\item{abstracts}{A dataframe including the minimum required variables: ID, documents, title, start.year.}

\item{label}{An optional label for the output. Default is an empty string.}

\item{lda.k.seq}{A sequence of integers specifying the number of topics for each model.}

\item{lda.prune.row}{Minimum number of unique terms required in a document for it to be included in the model. Default is 25.}

\item{lda.prune.col}{Minimum number of unique documents required for a term to be included in the model. Default is 5.}

\item{filter.terms}{An optional list of terms to filter the models.}
}
\value{
A list containing three elements:
\describe{
  \item{l.topics}{A list of topic models generated for different values of the lda.k.seq argument. Each element of the list corresponds to a different topic model, where the key is the number of topics used for that particular model.}
  \item{tidyterms}{A tibble containing cleaned and formatted tidy terms. See also 'topic.model.ready.tidy.terms' function.}
  \item{docterm}{A document-term matrix. Each row represents a document, each column represents a term, and values are the frequency of the corresponding term in the respective document. The document-term matrix is pruned to remove terms and documents that do not meet certain criteria before being used in the topic modeling process. See also 'prune.incidence' function.}
}
}
\description{
This function generates topic models from a dataset of abstracts. The dataset format is a slightly reformatted version of the export format from Zotero, see \link{abstracts}.
}
\examples{
# Load example abstracts data
data(abstracts)
# Set minimum requirements for abstract data
example(abstracts)
# Run function with default settings
documents.to.topics(abstracts)
}
\seealso{
\code{\link{prune.incidence}}, \code{\link{topic.model.ready.tidy.terms}}, \code{\link{docterm.to.lda}}.
}
