% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{map.supterms}
\alias{map.supterms}
\title{Map supplementary terms in the space of topics}
\usage{
map.supterms(
  result,
  docterm,
  dim = c(1, 2),
  label = "",
  one.plus = "+ unknown",
  one.minus = "- unknown",
  two.plus = "+ unknown",
  two.minus = "- unknown",
  three.plus = "+ unknown",
  three.minus = "- unknown",
  guess.labels = TRUE,
  repel.text.size = 3,
  browse = FALSE
)
}
\arguments{
\item{result}{An object of class 'soc.ca' containing the MCA results as returned from the 'mca.from.topics' function.}

\item{docterm}{A document-term matrix as returned from the 'documents.to.topics' function.}

\item{dim}{The dimensions in the order they are to be plotted. The first number defines the horizontal axis and the second number defines the vertical axis.}

\item{label}{An optional label for the output. Default is an empty string.}

\item{one.plus}{Label for the positive direction in the first dimension.}

\item{one.minus}{Label for the negative direction in the first dimension.}

\item{two.plus}{Label for the positive direction in the second dimension.}

\item{two.minus}{Label for the negative direction in the second dimension.}

\item{three.plus}{Label for the positive direction in the third dimension.}

\item{three.minus}{Label for the negative direction in the third dimension.}

\item{guess.labels}{Logical. If TRUE, the function will attempt to guess labels for each direction based on the most contributing categories.}

\item{repel.text.size}{Size of text to be repelled in the plots.}

\item{browse}{Logical. If TRUE, the browser is invoked for debugging purposes.}
}
\value{
A plot of the supplementary terms.
}
\description{
This is a plotting function to map supplementary terms in the space of topics.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
topics <- docterm.to.lda(tm_list[[3]])
result <- mca.from.topics(topics)
docterm <- tm_list[[3]]
# Run function with default settings
map.supterms(result, docterm)
}
