% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{map.supvar}
\alias{map.supvar}
\title{Map supplementary variables in the space of topics}
\usage{
map.supvar(
  result,
  docterm,
  vars,
  label = "",
  one.plus = "+ unknown",
  one.minus = "- unknown",
  two.plus = "+ unknown",
  two.minus = "- unknown",
  three.plus = "+ unknown",
  three.minus = "- unknown",
  guess.labels = TRUE,
  repel.text.size = 3,
  browse = FALSE
)
}
\arguments{
\item{result}{An object of class 'soc.ca' containing the MCA results as returned from the 'mca.from.topics' function.}

\item{docterm}{A document-term matrix as returned from the 'documents.to.topics' function.}

\item{vars}{A dataframe with two columns, namely ID and the supplementary variable with two categories (TRUE/FALSE).}

\item{label}{An optional label for the output. Default is an empty string.}

\item{one.plus}{Label for the positive direction in the first dimension.}

\item{one.minus}{Label for the negative direction in the first dimension.}

\item{two.plus}{Label for the positive direction in the second dimension.}

\item{two.minus}{Label for the negative direction in the second dimension.}

\item{three.plus}{Label for the positive direction in the third dimension.}

\item{three.minus}{Label for the negative direction in the third dimension.}

\item{guess.labels}{Logical. If TRUE, the function will attempt to guess labels for each direction based on the most contributing categories.}

\item{repel.text.size}{Size of text to be repelled in the plots.}

\item{browse}{Logical. If TRUE, the browser is invoked for debugging purposes.}
}
\value{
A plot displaying the risk ratios of terms according to a supplementary variable.
}
\description{
This is a plotting function to map supplementary variables in the space of topics. The function calculates and maps the risk ratios of terms according to a supplementary variable from the 'abstracts' or the 'authors' dataframe.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
topics <- docterm.to.lda(tm_list[[3]])
result <- mca.from.topics(topics)
docterm <- tm_list[[3]]
vars <- abstracts \%>\% mutate(var = Language \%in\% c("en")) \%>\% select(ID, var)
# vars <- authors \%>\% mutate(var = grepl("Bourdieu, P", Author)) \%>\% select(ID, var)
# Run function with default settings
map.supvar(result, docterm, vars)
}
