% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mca.from.topics}
\alias{mca.from.topics}
\title{Step 3. Make MCA from topics}
\usage{
mca.from.topics(topics, s = 0.01, small = FALSE)
}
\arguments{
\item{topics}{A dataframe containing the distribution of topics for each document as returned from the 'docterm.to.lda' function.}

\item{s}{A numeric value specifying the cut-off threshold for MCA analysis. Default is 0.01.}

\item{small}{Logical. If FALSE (the default), the function will retain additional matrices in the output for potential further analysis. If TRUE, additional matrices are excluded to save memory. This is useful when only the primary MCA results are needed and memory conservation is important.}
}
\value{
An object of class 'soc.ca' containing the MCA results. For more information on the different elements of the 'soc.ca' object, please refer to the documentation of the 'soc.ca' package.
}
\description{
This function performs an MCA based on the topics as returned from the 'docterm.to.lda' function.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
topics <- docterm.to.lda(tm_list[[3]])
# Run function with default settings
mca.from.topics(topics)
}
\seealso{
\code{\link{documents.to.topics}}, \code{\link{docterm.to.lda}}, \code{\link{map.topics}}, \code{\link{map.documents}}.
}
