% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mca.variance.plot}
\alias{mca.variance.plot}
\title{Plot MCA variance}
\usage{
mca.variance.plot(l.mca.seqs, n_topics = 100)
}
\arguments{
\item{l.mca.seqs}{A list of MCA results for each specified combination of cut-off and number of topics considered, as returned from the 'topics.to.mca' function.}

\item{n_topics}{A numeric value specifying the k number of topics intended for LDA. Default is 100.}
}
\value{
A ggplot of with the explained variance per cut off and number of topics
}
\description{
This function provides a plot of MCA explained variance for axes 1 and 2 according to different mca cut-off and number of topics considered.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
l.mca.seq <- topics.to.mca(tm_list[[1]])
# Run function with default settings
mca.variance.plot(l.mca.seq)
}
\seealso{
\code{\link{documents.to.topics}}, \code{\link{topics.to.mca}}
}
