% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prune.incidence}
\alias{prune.incidence}
\title{Prune incidence matrix}
\usage{
prune.incidence(
  incidence,
  min.col = nrow(incidence) * 0.05,
  min.row = 5,
  silent = FALSE
)
}
\arguments{
\item{incidence}{An incidence matrix to be pruned.}

\item{min.col}{Minimum number of nonzero elements (columns) required for retention. Default is 5 percent of the number of rows in the original matrix.}

\item{min.row}{Minimum number of nonzero elements (rows) required for retention. Default is 5.}

\item{silent}{Logical indicating whether to suppress output messages. Default is FALSE.}
}
\value{
A pruned incidence matrix.
}
\description{
This function prunes an incidence matrix by removing rows and columns with too few nonzero elements, based on specified thresholds. The pruning occurs in multiple iterations.
}
\examples{
# Example usage:
# incidence_matrix <- ...
# pruned_matrix <- prune.incidence(incidence_matrix)

}
\seealso{
This function is used within other functions. See \code{\link{documents.to.topics}} and \code{\link{mca.from.topics}}.
}
