% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sample.distinctive.terms}
\alias{sample.distinctive.terms}
\title{Sample Distinctive Terms}
\usage{
sample.distinctive.terms(result, docterm, s = 0.5)
}
\arguments{
\item{result}{An object of class 'soc.ca' containing the MCA results as returned from the 'mca.from.topics' function.}

\item{docterm}{A document-term matrix as returned from the 'documents.to.topics' function.}

\item{s}{A numeric value indicating the threshold for selecting distinctive terms. Default is 0.5.}
}
\value{
A tibble containing the sample of distinctive terms including frequency and coordinates.
}
\description{
This function returns the sample of the distinctive terms.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
topics <- docterm.to.lda(tm_list[[3]])
result <- mca.from.topics(topics)
# Run function with default settings
sample.distinctive.terms(result, tm_list[[3]])
}
\seealso{
\code{\link{documents.to.topics}}, \code{\link{docterm.to.lda}}, \code{\link{mca.from.topics}}, , \code{\link{supterms}}.
}
