% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{supterms.coord.and.risk.ratios}
\alias{supterms.coord.and.risk.ratios}
\title{Get coordinates for supplementary terms and risk ratios for supplementary variables}
\usage{
supterms.coord.and.risk.ratios(result, docterm, vars)
}
\arguments{
\item{result}{An object of class 'soc.ca' containing the MCA results as returned from the 'mca.from.topics' function.}

\item{docterm}{A document-term matrix as returned from the 'documents.to.topics' function.}

\item{vars}{A dataframe with two columns, namely ID and the supplementary variable.}
}
\value{
A tibble containing the coordinates and risk ratios for the supplementary variables.
}
\description{
This functions calculates the coordinates for supplementary terms and the risk ratios for supplementary variables. Supplementary variables are part of the abstracts dataframe or the authors dataframe.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
topics <- docterm.to.lda(tm_list[[3]])
result <- mca.from.topics(topics)
# Set vars argument
vars <- abstracts \%>\% mutate(var = Language \%in\% c("en")) \%>\% select(ID, var)
# vars <- authors \%>\% mutate(var = grepl("Bourdieu, P", Author)) \%>\% select(ID, var)
# Run function with default settings
supterms.coord.and.risk.ratios(result, tm_list[[3]], vars)
}
\seealso{
\code{\link{documents.to.topics}}, \code{\link{mca.from.topics}}, \code{\link{map.supvar}}
}
