% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{supterms}
\alias{supterms}
\title{Define supplementary terms}
\usage{
supterms(docterm, result, minimum.cases = result$n.ind * 0.05)
}
\arguments{
\item{docterm}{A document-term matrix as returned from the 'documents.to.topics' function.}

\item{result}{An object of class 'soc.ca' containing the MCA results as returned from the 'mca.from.topics' function.}

\item{minimum.cases}{The minimum number of cases for a term to be considered significant.}
}
\value{
A tibble with the supplementary terms associated to each document.
}
\description{
This function defines supplementary terms based on a minimum number of cases for a term to be considered significant.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
topics <- docterm.to.lda(tm_list[[3]])
result <- mca.from.topics(topics)
# Run function with default settings
supterms(tm_list[[3]], result)
}
\seealso{
\code{\link{documents.to.topics}}, \code{\link{docterm.to.lda}}, \code{\link{mca.from.topics}}, \code{\link{map.supterms}}
}
