% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{topic.model.ready.tidy.terms}
\alias{topic.model.ready.tidy.terms}
\title{Make tidyterms from raw text}
\usage{
topic.model.ready.tidy.terms(abstracts, long.label = FALSE, add.titles = TRUE)
}
\arguments{
\item{abstracts}{A dataframe including the following minimum required columns: ID, documents, title, start.year.}

\item{long.label}{Logical indicating whether to include long labels for word stems. Default is FALSE.}

\item{add.titles}{Logical indicating whether to add titles to the abstract text. Default is TRUE.}
}
\value{
A tibble with the tidy terms ready for topic modeling.
}
\description{
This function prepares raw text data for topic modeling by tokenizing, cleaning, and transforming it into a tidy format. It also calculates term frequency (tf), inverse document frequency (idf), and term frequency-inverse document frequency (tf-idf).
}
\examples{
# Load example abstracts data
data(abstracts)
# Set minimum requirements for abstract data
example(abstracts)
# Prepare tidy terms for topic modeling
topic.model.ready.tidy.terms(abstracts)
}
\seealso{
\code{\link{documents.to.topics}}
}
