% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{topic.pseudo.loglikelihood}
\alias{topic.pseudo.loglikelihood}
\title{Topic pseudo-loglikelihood}
\usage{
topic.pseudo.loglikelihood(
  l.topics,
  lda.k.seq = seq(from = 50, to = 200, by = 50),
  n_topics = 100
)
}
\arguments{
\item{l.topics}{A list of topic models as returned from 'documents.to.topics' function. Topics are generated for different values of the lda.k.seq argument in the 'documents.to.topics' function.}

\item{lda.k.seq}{A sequence of integers specifying the number of topics for each model. Must be the same as defined in the 'documents.to.topics' function.}

\item{n_topics}{A numeric value specifying the k number of topics intended for LDA. Default is 100.}
}
\value{
A plot of pseudo-loglikelihood for different numbers of topics
}
\description{
This function provides a plot of pseudo-loglikelihood for different k number of topics.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
# Run function with default settings
topic.pseudo.loglikelihood(tm_list[[1]])
}
\seealso{
\code{\link{documents.to.topics}}, \href{https://cran.r-project.org/web/packages/text2vec/index.html}{text2vec package}.
}
