% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{topics.to.mca}
\alias{topics.to.mca}
\title{Make a list of MCA from a list of topic models}
\usage{
topics.to.mca(
  l.topics,
  label = "",
  mca.cut.off.seq = seq(from = 0.01, to = 0.05, by = 0.01),
  mca.k.set = c(25, 50, 100, 150, 200)
)
}
\arguments{
\item{l.topics}{A list of topics. Refers to the first element of the list of three as returned from the 'documents.to.topics' function.}

\item{label}{An optional label for the output. Default is an empty string.}

\item{mca.cut.off.seq}{A sequence of values specifying the cut-off for MCA analysis.}

\item{mca.k.set}{A set of integers specifying the number of topics for the MCA.}
}
\value{
A list of MCA results for each specified combination of cut-off and number of topics considered.
}
\description{
This function performs multiple correspondence analysis (MCA) from a list of topic models as returned from the 'documents.to.topics' function.
}
\examples{
# Requirements
example(abstracts)
tm_list <- documents.to.topics(abstracts)
# Run function with default settings
topics.to.mca(tm_list[[1]])
}
\seealso{
\code{\link{documents.to.topics}}.
}
